+++ core/src/test/java/de/qabel/core/index/server/VerificationCodeEndpointTest.kt
@@ -1,4 +1,4 @@
-package de.qabel.core.index
+package de.qabel.core.index.server
 
 import de.qabel.core.index.server.IndexHTTPLocation
 import de.qabel.core.index.server.VerificationCodeEndpointImpl
